requirements.txt
Django==2.2.5
django-js-asset==1.2.2
django-mptt==0.10.0
django-rest-framework==0.1.0
django-tagging==0.4.6
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
mysqlclient==1.4.4
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.2
sqlparse==0.3.0
